pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ek-transcript"
version = "0.1.0"
description = "Speaker diarization transcription pipeline using AWS serverless"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "ekusiadadus" }
]
keywords = ["transcription", "speaker-diarization", "whisper", "aws", "lambda"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "boto3>=1.42.0",
    "botocore>=1.42.0",
    "python-dotenv>=1.2.0",
]

[project.optional-dependencies]
extract-audio = [
    # ffmpeg は subprocess で直接呼び出し（外部ライブラリ不要）
]
diarize = [
    "pyannote.audio>=4.0.2",
    # torch==2.8.0 は pyannote.audio が自動インストール
]
transcribe = [
    "faster-whisper>=1.2.0",
]
split = [
    # ffmpeg は subprocess で直接呼び出し（外部ライブラリ不要）
]
llm = [
    "openai>=2.9.0",
    "tenacity>=9.1.0",
]
google-auth = [
    "google-auth>=2.37.0",
    "google-auth-oauthlib>=1.2.0",
    "google-api-python-client>=2.160.0",
]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.0",
    "pytest-asyncio>=1.3.0",
    "moto[s3,secretsmanager]>=5.1.0",
    "ruff>=0.14.0",
    "mypy>=1.19.0",
    "black>=25.11.0",
    "isort>=7.0.0",
    "pre-commit>=4.5.0",
    "types-boto3>=1.0.0",
]
all = [
    "ek-transcript[extract-audio,diarize,transcribe,split,llm,google-auth,dev]",
]

[project.urls]
Homepage = "https://github.com/ekusiadadus/ek-transcript"
Repository = "https://github.com/ekusiadadus/ek-transcript"

[tool.hatch.build.targets.wheel]
packages = ["lambdas"]

[tool.pytest.ini_options]
testpaths = ["tests", "lambdas"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["lambdas"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 80

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "S",   # flake8-bandit (security)
]
ignore = [
    "E501",  # line too long (handled by black)
    "S101",  # use of assert (ok in tests)
    "S108",  # /tmp usage (required for Lambda)
    "S603",  # subprocess call (trusted ffmpeg command in Lambda)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]
"**/tests/*" = ["S101", "B017"]

[tool.ruff.lint.isort]
known-first-party = ["lambdas"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[[tool.mypy.overrides]]
module = "test_handler"
disallow_untyped_defs = false
warn_return_any = false
disable_error_code = ["misc"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["lambdas"]
